tests: Use realview-eb instead of realview-pbx-a9
QEMU implements 4 different "realview" machine types:

  $ qemu-system-aarch64 -machine help 2>&1 | grep realview
  realview-eb          ARM RealView Emulation Baseboard (ARM926EJ-S)
  realview-eb-mpcore   ARM RealView Emulation Baseboard (ARM11MPCore)
  realview-pb-a8       ARM RealView Platform Baseboard for Cortex-A8
  realview-pbx-a9      ARM RealView Platform Baseboard Explore for Cortex-A9

Of these, only the -eb variants support PCI devices and are
thus relevant when it comes to USB controllers.

Our logic treats all these machine types the same, which is
incorrect. An upcoming commit will fix the issue; in
preparation for that, make some adjustments to the test suite.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>
Reviewed-by: Peter Krempa <pkrempa@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>